+++ AdventOfCode2019/Shared/DummySignal.cs
@@ -5,12 +5,12 @@ namespace AdventOfCode2019.Shared
 {
     public class DummySignal : Signal
     {
-        public void Output(int value)
+        public void Output(long value)
         {
             throw new Exception("Dummy signal should never be called");
         }
 
-        public Task<int> Input()
+        public Task<long> Input()
         {
             throw new Exception("Dummy signal should never be called");
         }
